+++ docs/CHANGELOG.md
@@ -6,6 +6,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] 
 
+## [0.6.0] 2017-06-07
+### Added
+- Syntax highlighting for code examples
+- Copy to clipboard icons for examples
+### Fixed
+- added missing cost indicators to methods
+
+## [0.5.2] 2017-06-07
+### Fixed
+- added missing Authenticated denotations with icons to endpoints
+
+## [0.5.1] 2017-06-07
+### Fixed
+- Made the $api_root and $version variables from yaml render to the constants in the prebuild step
+
+## [0.5.0] 2017-06-06
+### Added
+- cURL guide #1878
+### Fixed
+- updated duplicate base version vars to use API_VERSION, prebuild script uses environment var if present
+- include original /reference redirect to the new /v4/introduction #2019
+### Changed
+- updated Introduction section #2008
+
 ## [0.4.0] 2017-06-05
 ### Added
 - /v4 to the base of all urls